don't override renamed assets

+++ copy-assets.js
@@ -3,7 +3,7 @@ import fs from "fs";
 // copy folder ./src/assets to ./dist/assets
 const reader = fs.opendirSync("./src/assets");
 for await (const dirent of reader) {
-	if (dirent.isFile()) {
+	if (dirent.isFile() && !dirent.name.endsWith(".ts")) {
 		fs.copyFileSync(
 			`./src/assets/${dirent.name}`,
 			`./dist/assets/${dirent.name}`
